ADDED BOTTOM NAVBAR WORKING
