+++ @narative/gatsby-theme-novela/src/components/BottomNavBar/index.ts
@@ -1,0 +1,1 @@
+export { default } from '@components/BottomNavBar/BottomNavBar';
